requirements.txt
altgraph==0.17
appdirs==1.4.4
black==21.7b0
click==8.0.1
colorama==0.4.4
future==0.18.2
mypy-extensions==0.4.3
pathspec==0.9.0
pefile==2021.5.24
Pillow==8.3.1
pyinstaller-hooks-contrib==2021.2
PySimpleGUI==4.46.0
pysimplegui-exemaker==1.3.0
pywin32-ctypes==0.2.0
regex==2021.8.3
tomli==1.2.1
tqdm==4.62.1
